% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.CalcFunctions.R
\name{akfin.calcBycatchABs}
\alias{akfin.calcBycatchABs}
\title{Calculate estimated total bycatch abundance and biomass of Tanner crab in groundfish fisheries}
\usage{
akfin.calcBycatchABs(
  maxYear = 2019,
  fnHIS = "GroundfishFisheries.HistoricalBycatchBiomass.TannerCrab.csv",
  fnCAS = "FromAKFIN.TannerCrab.BycatchEstimates.CAS.csv",
  fnCIA = "FromAKFIN.TannerCrab.BycatchEstimates.CIA.csv",
  verbosity = 0
)
}
\arguments{
\item{maxYear}{\itemize{
\item max fishery year to extract
}}

\item{fnHIS}{\itemize{
\item filename for bycatch estimates from "historical data" (pre-CAS)
}}

\item{fnCAS}{\itemize{
\item filename for bycatch estimates from Catch Accounting System
}}

\item{fnCIA}{\itemize{
\item filename for bycatch estimates from Catch-In_Areas database
}}

\item{verbosity}{\itemize{
\item integer flag for diagnostic printout (0,1,2)
}}
}
\value{
tibble with columns
\itemize{
\item fishery
\item area
\item year
\item expFactor - expansion factor
\item sex - "male", "female", "undetermined", "hermaphrodite", or NA
\item variable - "count", "weight", "abundance", or "biomass"
\item value - value of associated variable
\item type - "observed" or "expanded"
\item units - "ones", "thousands", "kg" or "t"
}
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
Historical (foreign, joint veture) data runs 1973-1990. CAS data is used for 1991-2008. CIA data is used for
2009+.
}
