% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.calcCatchTables.R
\name{akfin.calcCatchTables}
\alias{akfin.calcCatchTables}
\title{Calculate bycatch tables for Tanner crab in groundfish fisheries}
\usage{
akfin.calcCatchTables(
  maxYear = 2019,
  dirData = "~/Work/StockAssessments-Crab/Data/Fisheries/Fishery.AKFIN/Current",
  fnHisABs = "GroundfishFisheries.HistoricalABs.TannerCrab.csv",
  fnCAS = "FromAKFIN.TannerCrab.BycatchEstimates.CAS.csv",
  fnCIA = "FromAKFIN.TannerCrab.BycatchEstimates.CIA.csv",
  fnHisZCs = "GroundfishFisheries.HistoricalZCs.LongFormat.TannerCrab.csv",
  fnNORPAC = "FromAKFIN.TannerCrab.norpac_length_report.csv",
  verbosity = 0
)
}
\arguments{
\item{maxYear}{- max fishery year to extract}

\item{dirData}{- path to "current" AKFIN data directory}

\item{fnHisABs}{- filename for bycatch estimates (abundance, biomass) from "historical data" (pre-CAS)}

\item{fnCAS}{- filename for bycatch estimates (abundance, biomass) from Catch Accounting System}

\item{fnCIA}{- filename for bycatch estimates (abundance, biomass) from Catch-In_Areas database}

\item{fnHisZCs}{- filename for size composition estimates (numbers of crab) from "historical data" (pre-CAS)}

\item{fnNORPAC}{- filename for size data (numbers of crab) from a NORPAC length report file}

\item{verbosity}{- integer flag for diagnostic printout (0,1,2)}
}
\value{
list with the following tibbles: \cr
\itemize{
\item{dfrABs.YGAT - total bycatch abundance, biomass by YGAT}
\item{dfrSSs.YGAX - "raw" sample sizes for size compositions by YGAX}
\item{dfrZCs.onYGAXZ - "raw" size compositions by YGAXZ}
\item{dfrZCs.tnYGAXZ - total catch (if possible) size compositions by YGAXZ}
}
}
\description{
Function to calculate bycatch tables for Tanner crab in groundfish fisheries.
}
\details{
Historical (foreign, joint venture) biomass data runs 1973-1990 (no abundance data 
is available during this time period). CAS data is used for 1991-2008. CIA data is used for
2009+. Historical size composition data runs 1973-1990. NORPAC size composition data starts
(effectively) in 1986, but is used here starting in 1991.
}
\note{
Abundance is in number of crab, biomass in kg.
}
