% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.CalcFunctions.R
\name{akfin.calcInputSSsForTCSAM}
\alias{akfin.calcInputSSsForTCSAM}
\title{Calculate estimated total bycatch abundance and biomass of Tanner crab in groundfish fisheries}
\usage{
akfin.calcInputSSsForTCSAM(dfrZCs, ss_scl, sss_max = 200)
}
\arguments{
\item{dfrZCs}{- filename for size composition estimates (numbers of crab) from "historical data" (pre-CAS)}

\item{ss_scl}{- filename for size data (numbers of crab) from a NORPAC length report file}

\item{ss_max}{- integer flag for diagnostic printout (0,1,2)}
}
\value{
tibble with columns
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
Historical (foreign, joint veture) data runs 1973-1990. CAS data is used for 1991-2008. CIA data is used for
2009+.
}
