% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.CalcFunctions.R
\name{akfin.calcUnscaledZCs}
\alias{akfin.calcUnscaledZCs}
\title{Calculate estimated total bycatch abundance and biomass of Tanner crab in groundfish fisheries}
\usage{
akfin.calcUnscaledZCs(
  maxYear = 2019,
  fnHisZCs = NULL,
  fnNORPAC = NULL,
  id.facs = c("year", "gear", "area", "sex"),
  cutpts = seq(25, 185, 5),
  truncate.low = TRUE,
  truncate.high = FALSE
)
}
\arguments{
\item{maxYear}{- max fishery year to extract}

\item{fnHisZCs}{- filename for size composition estimates (numbers of crab) from "historical data" (pre-CAS)}

\item{fnNORPAC}{- filename for size data (numbers of crab) from a NORPAC length report file}

\item{verbosity}{- integer flag for diagnostic printout (0,1,2)}
}
\value{
tibble with columns
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
Historical (foreign, joint veture) data runs 1973-1990. CAS data is used for 1991-2008. CIA data is used for
2009+.
}
