% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.ConvertGearCodeToType}
\alias{akfin.ConvertGearCodeToType}
\title{Create dataframe describing gear conversions}
\usage{
akfin.ConvertGearCodeToType(x)
}
\value{
character vector with same length as x
}
\description{
Function to create dataframe describing gear conversions.
}
\details{
None.
}
