% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.ConvertGearDescToType}
\alias{akfin.ConvertGearDescToType}
\title{Convert gear descriptions to gear types}
\usage{
akfin.ConvertGearDescToType(x)
}
\arguments{
\item{x}{- vector of gear descriptions to convert to gear types}
}
\value{
dataframe with columns
}
\description{
Function to convert gear descriptions to gear types.
}
\details{
None.
}
