% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.ConvertSexCodes}
\alias{akfin.ConvertSexCodes}
\title{Convert sex codes to descriptions}
\usage{
akfin.ConvertSexCodes(x)
}
\arguments{
\item{x}{- vector of sex codes to convert}
}
\value{
character vector with same length as x
}
\description{
Function to convert sex codes to descriptions.
}
\details{
None.
}
