% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.GetGearConversions}
\alias{akfin.GetGearConversions}
\title{Create dataframe describing gear conversions}
\usage{
akfin.GetGearConversions()
}
\value{
dataframe with columns
}
\description{
Function to create dataframe describing gear conversions.
}
\details{
None.
}
