% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.GetSexConversions}
\alias{akfin.GetSexConversions}
\title{Create dataframe describing sex conversions}
\usage{
akfin.GetSexConversions()
}
\value{
dataframe with columns
}
\description{
Function to create dataframe describing sex conversions.
}
\details{
None.
}
