% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.CalcFunctions.R
\name{akfin.ScaleZCs}
\alias{akfin.ScaleZCs}
\title{Calculate estimated total bycatch abundance and biomass of Tanner crab in groundfish fisheries}
\usage{
akfin.ScaleZCs(dfrABs.YGAT, dfrZCs.onYGAXZ)
}
\arguments{
\item{dfrABs.YGAT}{- dataframe with total abundance by YGAT}

\item{dfrZCs.onYGAXZ}{- dataframe with "raw" number size compositions by YGAX}
}
\value{
tibble with columns
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
Historical (foreign, joint veture) data runs 1973-1990. CAS data is used for 1991-2008. CIA data is used for
2009+.
}
