% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfinConvert_ADFGStatAreas-LatLons.R
\name{akfinConvert_ADFGStatAreasToLatLons}
\alias{akfinConvert_ADFGStatAreasToLatLons}
\title{Convert ADFG stat areas to lat-lon coordinates}
\usage{
akfinConvert_ADFGStatAreasToLatLons(stat_areas)
}
\arguments{
\item{stat_areas}{: character vector of ADFG stat areas}
}
\value{
a tibble with columns "lat", "lon" representing the center of each stat area.
}
\description{
Function to convert ADFG stat areas to lat-lon coordinates.
}
\details{
ADFG stat areas are coded AABBCC, where\cr
lon = 100+AA W longitude \cr
and \cr
lat = BB+CC/60 N latitude \cr
in decimal degrees.

AABBCC represents the lower right corner
of the stat area.
}
