% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfinConvert_ADFGStatAreas-LatLons.R
\name{akfinConvert_LatLonsToADFGStatAreas}
\alias{akfinConvert_LatLonsToADFGStatAreas}
\title{Convert lat-lon coordinates to ADFG stat areas}
\usage{
akfinConvert_LatLonsToADFGStatAreas(dfr, latCol = "lat", lonCol = "lon")
}
\arguments{
\item{dfr}{: dataframe with lat, lon coordinates}

\item{latCol}{: name of column with latitudes}

\item{lonCol}{: name of column with longitudes}
}
\value{
a character vector with the ADFG stat area corresponding to each set
of lat-lon coordinates.
}
\description{
Function to convert lat-lon coordinates to ADFG stat areas.
}
\details{
ADFG stat areas are coded AABBCC, where\cr
lon = 100+AA W longitude \cr
and \cr
lat = BB+CC/60 N latitude \cr
in decimal degrees.

AABBCC represents the lower right corner
of the stat area.
}
