% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfinConvert_SexCodes}
\alias{akfinConvert_SexCodes}
\title{Convert sex codes to descriptions}
\usage{
akfinConvert_SexCodes(x)
}
\arguments{
\item{x}{: vector of sex codes to convert}
}
\value{
character vector with same length as x
}
\description{
Function to convert sex codes to descriptions.
}
\details{
Uses conversions from \code{code} to \code{sex}
defined by \code{\link[=akfinGet_SexConversions]{akfinGet_SexConversions()}}.
}
