% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfinGet_GearConversions}
\alias{akfinGet_GearConversions}
\title{Create dataframe describing AKFIN gear conversions}
\usage{
akfinGet_GearConversions()
}
\value{
dataframe with columns 'type', 'code', and 'description'
}
\description{
Function to create dataframe describing gear conversions.\cr

"type" ("code") can be \cr
"trawl" ("TRW", "PTR" or "NPT") or \cr
"fixed" ("HAL","JIG","POT") \cr

"code" ("description") can be \cr
"TRW" ("TRAWL"),\cr
"PTR" ("PELAGIC" or "PAIR TRAWL"),\cr
"NPT" ("NON PELAGIC"),\cr
"HAL" ("LONGLINER"),\cr
"POT" ("POT OR TRAP"),\cr
"JIG" ("JIG")\cr
}
\details{
None.
}
