% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfinGet_SexConversions}
\alias{akfinGet_SexConversions}
\title{Create dataframe describing sex conversions}
\usage{
akfinGet_SexConversions()
}
\value{
dataframe with columns 'sex' and 'code'.
}
\description{
Function to create dataframe describing sex conversions.
}
\details{
None.
}
