% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.ReadCSVs.R
\name{akfinRead_NorpacLengthReport}
\alias{akfinRead_NorpacLengthReport}
\title{Read a csv file from AKFIN with NORPAC length report data}
\usage{
akfinRead_NorpacLengthReport(fn)
}
\arguments{
\item{fn}{\itemize{
\item filename for size composition data from the NORPAC database
}}
}
\value{
tibble with columns "year","gear","area","sex","size",and "N",
"gear code","lat","lon",and "adfg stat area". "gear" here represents the
gear "type" from the Conversion functions whereas "gear code" represents the
gear "code" from those functions.

Units for N are counts (one's) of sampled crab.
}
\description{
Function to read a csv file from AKFIN with NORPAC length report data.
}
\details{
Valid NORPAC data begins in 1986, but is typically used starting in 1991, to the present.
}
