% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfinWrite_TCSAMInputFile.R
\name{akfinWrite_TCSAMInputFile}
\alias{akfinWrite_TCSAMInputFile}
\title{Write AKFIN data for crab bycatch to a TCSAM input file}
\usage{
akfinWrite_TCSAMInputFile(
  fishery = NULL,
  gear = "all",
  fn = "",
  closed = NULL,
  dfrTC_ABs = NULL,
  dfrTC_ZCs = NULL,
  dfrTC_SSs = NULL,
  dfrEffort = NULL,
  cutpts = NULL,
  ss_factor = 1,
  ss_max = 200,
  likeTC = "NORM2",
  cvTC = 0.2,
  minErrTCa = 2000,
  minErrTCb = 2000,
  unitsBiomass = "THOUSANDS_MT"
)
}
\arguments{
\item{fishery}{- TCSAM fishery name}

\item{gear}{- gear type}

\item{fn}{- output file name (default writes to stdout)}

\item{closed}{- vector of years when fishery was closed}

\item{dfrTC_ABs}{- dataframe with total bycatch abundance/biomass data}

\item{dfrTC_ZCs}{- dataframe with total bycatch size composition data}

\item{dfrTC_SSs}{- dataframe with total bycatch sample size information}

\item{cutpts}{- vector with cutpts used for size comps}

\item{ss_factor}{- sample size scaling factor (from retained catch analysis)}

\item{ss_max}{- max for sample sizes (default=200)}

\item{likeTC}{- likelihood type for total bycatch ABs ("NORM2","NORMAL", or "LOGNORMAL")}

\item{cvTC}{- cv for total bycatch above min error (default=0.20)}

\item{minErrTCa}{- min error in total bycatch abundance (default=2000)}

\item{minErrTCb}{- min error in total bycatch biomass   (default=2000 kg)}

\item{unitsBiomass}{- units for output biomass ("THOUSANDS_MT" [default],"MILLIONS_LBS","KG")}
}
\value{
null
}
\description{
Function to write AKFIN data for crab bycatch to a TCSAM input file.
}
\details{
None
}
