% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.AllConversions.R
\name{akfin.ConvertGearDescToType}
\alias{akfin.ConvertGearDescToType}
\title{Convert gear descriptions to gear types}
\usage{
akfin.ConvertGearDescToType(x)
}
\arguments{
\item{x}{- vector of gear descriptions to convert to gear types}
}
\value{
character vector with same length as x
}
\description{
Function to convert gear descriptions to gear types.
}
\details{
Uses conversions from \code{description} to \code{type} 
defined by \code{akfin.GetGearConversions}.
}
