% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.ReadCSVs.R
\name{akfin.ReadCSV.CIA}
\alias{akfin.ReadCSV.CIA}
\title{Read a csv file from AKFIN with CIA (Catch-in-Areas) data}
\usage{
akfin.ReadCSV.CIA(fn, minYear = 2009)
}
\arguments{
\item{fn}{- filename for bycatch estimates from the Catch-in-Areas database}

\item{minYear}{- min fishery year to extract}
}
\value{
tibble with columns "year","target","gear code","adfg stat area","nmfs stat area",
"haul count","biomass","number","conf flag","vessel count", and "gear". Units for biomass
and number are kg and one's (i.e., unscaled number of crab).
}
\description{
Function to read a csv file from AKFIN with CIA (Catch-in-Areas) data.
}
\details{
CIA data begins in 2009.
}
