% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfin.CalcFunctions.R
\name{akfin.ScaleZCs}
\alias{akfin.ScaleZCs}
\title{Calculate estimated total bycatch abundance-scaled size comps of Tanner crab in groundfish fisheries}
\usage{
akfin.ScaleZCs(
  dfrABs.YGAT,
  dfrZCs.onYGAXZ,
  cutpts = seq(25, 185, 5),
  truncate.low = TRUE,
  truncate.high = FALSE
)
}
\arguments{
\item{dfrABs.YGAT}{\itemize{
\item dataframe with total abundance by YGAT
}}

\item{dfrZCs.onYGAXZ}{\itemize{
\item dataframe with "raw" number size compositions by YGAX
}}

\item{cutpts}{\itemize{
\item cutpts for scale size compositions (default is 25->185 by 5)
}}

\item{truncate.low}{-flag to truncate counts/abundance below the minimum size or include in first size bin (FALSE)}

\item{truncate.high}{\itemize{
\item flag to truncate counts/abundance above the maximum size (TRUE) or include in final size bin (FALSE)
}}
}
\value{
tibble with columns 'year','gear','area','sex','maturity','shell condition','size','N'
}
\description{
Function to calculate estimated total bycatch abundance-scaled size comps of Tanner crab in groundfish fisheries.
}
\details{
Estimated total annual bycatch abundance from the CAS and CIA databases (as a tibble from
\code{akfin.calcBycatchABs}) is used to scale unscaled size compositions (a tibble from \code{akfin.calcUnscaledZCs})
for years in which such exists. For years in which total bycatch abundance is not available, the size
compositions are not scaled.
}
