% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akfinWrite_TCSAMInputFile.R
\name{akfinWrite_TCSAMInputFile}
\alias{akfinWrite_TCSAMInputFile}
\title{Write AKFIN data for crab bycatch to a TCSAM input file}
\usage{
akfinWrite_TCSAMInputFile(
  fishery = NULL,
  fn = "Data.Fishery.AKFIN.inp",
  lstAbd = NULL,
  lstBio = NULL,
  lstZCs = NULL
)
}
\arguments{
\item{fishery}{- TCSAM fishery name}

\item{fn}{- output file name (writes to stdout if NULL or empty)}

\item{lstAbd}{- NULL, or list as returned by \link[tcsamFunctions]{inputList_AggregateCatchData} with type "ABUNDANCE"}

\item{lstBio}{- NULL, or list as returned by \link[tcsamFunctions]{inputList_AggregateCatchData} with type "BIOMASS"}

\item{lstZCs}{- NULL, or list as returned by \link[tcsamFunctions]{inputList_SizeCompsData}}
}
\value{
nothing
}
\description{
Function to write AKFIN data for crab bycatch to a TCSAM input file.
}
\details{
Calls \link[tcsamFunctions]{writeInputFile_FleetData} with \code{type}="FISHERY" (and lstIC, lstRC, lstDC, and lstEff set to NULL).
}
